+++ packages/codecs/src/feed.ts
@@ -1,5 +1,5 @@
 import { type Context, Type, type } from 'codeco'
-import { commitIdAsString, StreamMetadata } from './stream.js'
+import { commitIdAsString, EventTypeAsNumber, StreamMetadata } from './stream.js'
 
 export const JsonAsString = new Type<unknown, string, string>(
   'JSON-as-string',
@@ -18,4 +18,5 @@ export const AggregationDocument = type({
   commitId: commitIdAsString,
   content: JsonAsString,
   metadata: StreamMetadata,
+  eventType: EventTypeAsNumber,
 })