+++ packages/indexing/src/database-index-api.ts
@@ -28,11 +28,10 @@ export interface IndexStreamArgs {
   readonly streamID: StreamID
   readonly model: StreamID
   readonly controller: string
-  readonly streamContent: Record<string, any>
+  readonly streamContent: Record<string, any> | null
   readonly tip: CID
   readonly lastAnchor: Date | null
   readonly firstAnchor: Date | null
-  readonly shouldIndex?: boolean
 }
 
 /**
@@ -208,25 +207,18 @@ export abstract class DatabaseIndexApi<DateType = Date | number> {
   }
 
   /**
-   * This method inserts the stream if it is not present in the index, updates
-   * the 'content' if the stream already exists in the index, or deletes the
-   * stream from the index if the 'shouldIndex' arg is set to false.
+   * This method inserts the stream if it is not present in the index, or updates
+   * the 'content' if the stream already exists in the index.
    * @param indexingArgs
    */
   async indexStream(
     indexingArgs: IndexStreamArgs & { createdAt?: Date; updatedAt?: Date }
   ): Promise<void> {
     const tableName = asTableName(indexingArgs.model)
-    if (indexingArgs.shouldIndex === false) {
-      await this.dbConnection(tableName)
-        .where('stream_id', indexingArgs.streamID.toString())
-        .delete()
-      return
-    }
     const indexedData = this.getIndexedData(indexingArgs) as Record<string, unknown>
     const relations = this.modelRelations.get(indexingArgs.model.toString()) ?? []
     for (const relation of relations) {
-      indexedData[addColumnPrefix(relation)] = indexingArgs.streamContent[relation]
+      indexedData[addColumnPrefix(relation)] = indexingArgs.streamContent?.[relation]
     }
     const toMerge = cloneDeep(indexedData)
     delete toMerge['created_at']
@@ -450,7 +442,7 @@ export class PostgresIndexApi extends DatabaseIndexApi<Date> {
     return {
       stream_id: indexingArgs.streamID.toString(),
       controller_did: indexingArgs.controller.toString(),
-      stream_content: indexingArgs.streamContent,
+      stream_content: indexingArgs.streamContent ?? {},
       tip: indexingArgs.tip.toString(),
       last_anchored_at: indexingArgs.lastAnchor,
       first_anchored_at: indexingArgs.firstAnchor,
@@ -501,7 +493,7 @@ export class SqliteIndexApi extends DatabaseIndexApi<number> {
     return {
       stream_id: indexingArgs.streamID.toString(),
       controller_did: indexingArgs.controller.toString(),
-      stream_content: JSON.stringify(indexingArgs.streamContent),
+      stream_content: JSON.stringify(indexingArgs.streamContent ?? {}),
       tip: indexingArgs.tip.toString(),
       last_anchored_at: asTimestamp(indexingArgs.lastAnchor),
       first_anchored_at: asTimestamp(indexingArgs.firstAnchor),