+++ packages/stream-model-instance-handler/src/model-instance-document-handler.ts
@@ -7,7 +7,7 @@ import {
 import {
   AnchorStatus,
   CommitData,
-  CommitType,
+  EventType,
   SignatureStatus,
   SignatureUtils,
   StreamConstructor,
@@ -129,7 +129,7 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
       metadata,
       signature: SignatureStatus.SIGNED,
       anchorStatus: AnchorStatus.NOT_REQUESTED,
-      log: [StreamUtils.commitDataToLogEntry(commitData, CommitType.GENESIS)],
+      log: [StreamUtils.commitDataToLogEntry(commitData, EventType.INIT)],
     }
   }
 
@@ -163,18 +163,11 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
     )
 
     if (payload.header) {
-      const { shouldIndex, ...others } = payload.header
-      const otherKeys = Object.keys(others)
-      if (otherKeys.length) {
-        throw new Error(
-          `Unsupported metadata changes for ModelInstanceDocument Stream ${streamId}: ${otherKeys.join(
-            ','
-          )}. Only the shouldIndex argument can be changed.`
-        )
-      }
-      if (shouldIndex != null) {
-        state.metadata.shouldIndex = shouldIndex
-      }
+      throw new Error(
+        `Updating metadata for ModelInstanceDocument Streams is not allowed.  Tried to change metadata for Stream ${streamId} from ${JSON.stringify(
+          state.metadata
+        )} to ${JSON.stringify(payload.header)}\``
+      )
     }
 
     const oldContent = state.content ?? {}
@@ -190,7 +183,7 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
     state.signature = SignatureStatus.SIGNED
     state.anchorStatus = AnchorStatus.NOT_REQUESTED
     state.content = newContent
-    state.log.push(StreamUtils.commitDataToLogEntry(commitData, CommitType.SIGNED))
+    state.log.push(StreamUtils.commitDataToLogEntry(commitData, EventType.DATA))
 
     return state
   }
@@ -247,7 +240,11 @@ export class ModelInstanceDocumentHandler implements StreamHandler<ModelInstance
 
     validateContentLength(content)
 
-    this._schemaValidator.validateSchema(content, model.content.schema, model.commitId.toString())
+    await this._schemaValidator.validateSchema(
+      content,
+      model.content.schema,
+      model.commitId.toString()
+    )
 
     // Now validate the relations
     await this._validateRelationsContent(ceramic, model, content)